+++ src/components/friends/async-render/render.ts
@@ -6,7 +6,6 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
-import { beforeMountHooks } from 'core/component/const';
 import { restart, deferRestart } from 'core/component/render/daemon';
 
 import type Friend from 'components/friends/friend';
@@ -61,7 +60,7 @@ export function waitForceRender(
 	elementToDrop?: string | ((ctx: Friend['component']) => CanPromise<CanUndef<string | Element>>)
 ): () => CanPromise<boolean> {
 	return () => {
-		if (beforeMountHooks[this.hook] != null) {
+		if (!this.ctx.renderedOnce) {
 			return true;
 		}