+++ src/components/super/i-static-page/modules/theme/system-theme-extractor/index.ts
@@ -1,0 +1,9 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+export * from 'components/super/i-static-page/modules/theme/system-theme-extractor/interface';