+++ src/core/component/directives/ref/index.ts
@@ -24,6 +24,10 @@ import type { DirectiveOptions } from 'core/component/directives/ref/interface';
 export * from 'core/component/directives/ref/const';
 export * from 'core/component/directives/ref/interface';
 
+//#if runtime has dummyComponents
+import('core/component/directives/ref/test/b-directives-ref-dummy');
+//#endif
+
 ComponentEngine.directive('ref', {
 	mounted: updateRef,
 	updated: updateRef
@@ -108,7 +112,7 @@ function updateRef(el: Element | ComponentElement, opts: DirectiveOptions, vnode
 			refVal = getRefVal();
 
 		let
-			ref;
+			ref: unknown;
 
 		if (Object.isArray(refVal)) {
 			if (key != null) {