+++ src/components/friends/async-render/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v4.0.0-beta.67 (2024-02-26)
+
+#### :bug: Bug Fix
+
+* Fixed synchronous render for functional components in `waitForceRender`
+
 ## v4.0.0-beta.46 (2024-01-11)
 
 #### :bug: Bug Fix